feat: creates cast in dynamic types

+++ lib/features/dashboard/data/models/multi_winner_year_model.dart
@@ -8,8 +8,8 @@ class MultiWinnerYearModel extends MultiWinnerYear {
 
   factory MultiWinnerYearModel.fromJson(Map<String, dynamic> json) {
     return MultiWinnerYearModel(
-      year: json['year'],
-      wins: json['winnerCount'],
+      year: json['year'] as int,
+      wins: json['winnerCount'] as int,
     );
   }
 }
